test(web): split page route test modules [OPE-296]
