+++ crates/opengoose-web/src/routes/pages/tests/catalog/mod.rs
@@ -1,0 +1,10 @@
+mod listing;
+mod schedules;
+mod teams;
+mod triggers;
+
+pub(super) use super::super::catalog::{
+    AgentQuery, RunQuery, ScheduleActionForm, ScheduleQuery, SessionQuery, TeamSaveForm,
+    TriggerActionForm, TriggerQuery, WorkflowQuery, agents, queue, runs, schedule_action,
+    schedules, sessions, team_save, trigger_action, triggers, workflows,
+};
