+++ crates/opengoose-web/src/routes/pages/tests/mod.rs
@@ -1,0 +1,5 @@
+mod catalog;
+mod dashboard;
+mod remote_agents;
+mod router;
+mod support;
